refactor(nexus-web): Chat UI module Phase 2 list chrome
Move ChatSection into the chat module with semantic CSS for conversation list chrome, keeping shell imports as thin re-exports.
